trying to make cells update by the grid, trying to work out rules for falling

+++ src/cell_manager.hpp
@@ -22,7 +22,7 @@ public:
 
     std::vector<sf::Vertex>* vertices;
 
-    int cellSize;
+    float cellSize;
 
     int cellColorVariance;
 };
